remove all init statements
