+++ site/ts/ta-grading-keymap-init.ts
@@ -1,0 +1,11 @@
+window.Twig.twig({
+    id: 'HotkeyList',
+    href: '/templates/grading/settings/HotkeyList.twig',
+    async: true,
+});
+
+window.Twig.twig({
+    id: 'GeneralSettingList',
+    href: '/templates/grading/settings/GeneralSettingList.twig',
+    async: true,
+});
